change: move `Room` implementations to separate files
